# gateway/api/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorResponse:
    properties:
      error:
        example: access denied
        type: string
    type: object
...
  user.CreateUserRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      department:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        type: string
      phone:
        type: string
      position:
        type: string
      role:
        type: string
    type: object
  user.ListUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/user.UserResponse'
        type: array
    type: object
  user.ManagedUserResponse:
    properties:
      email:
        type: string
      profile:
        $ref: '#/definitions/user.UserResponse'
      role:
        type: string
    type: object
  user.UpdateUserRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      department:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      position:
        type: string
    type: object
  user.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      birth_date:
        type: string
      created_at:
        type: string
      department:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      phone:
        type: string
      position:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:28080
info:
  contact: {}
  description: Public REST API exposed by the gateway. Internal microservices communicate
    only over gRPC and are intentionally omitted from this specification.
  title: Project Management Gateway API
  version: "1.0"
paths:
...
  /api/users:
    get:
      description: Returns all user profiles. This endpoint is available only to administrators.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.ListUsersResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user profiles
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user or administrator account. This endpoint is available
        only to administrators.
      parameters:
      - description: Managed user payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.ManagedUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a managed user
      tags:
      - users
  /api/users/{id}:
    delete:
      description: Deletes a user in both auth-service and user-service. This endpoint
        is available only to administrators.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
    get:
      description: Returns a user profile by ID. Administrators can read any profile,
        regular users can read only their own profile.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates a user profile. Administrators can update any profile,
        regular users can update only their own profile.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Profile update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT access token. Use the `Bearer {token}` format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"